requirements.txt
aiogram==3.13.1
aiogram-calendar==0.5.0
aiohttp==3.10.11
APScheduler==3.10.4
tortoise-orm==0.21.6
aerich==0.7.2
asyncpg==0.29.0
python-dotenv==1.0.0
cryptography==42.0.8
aiomysql==0.2.0
